pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vrc-live-caption"
version = "0.1.0"
description = "Real-time speech caption tool for VRChat via OSC."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-osc>=1.9.3",
    "pyside6>=6.11.0",
    "rich>=14.3.3",
    "sounddevice>=0.5.5",
    "typer>=0.24.1",
    "websockets>=16.0",
    "deepl>=1.29.0",
    "google-cloud-translate>=3.25.0",
    "uharfbuzz>=0.53.6",
    "uniseg>=0.10.1",
    "fonttools>=4.62.1",
]

[project.scripts]
vrc-live-caption = "vrc_live_caption.cli:main"

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
    "ty>=0.0.26",
]

[project.optional-dependencies]
local-cpu = [
    "funasr[llm]>=1.3.1",
    "ilt-editdistance>=0.8.2",
    "sentencepiece>=0.2.1",
    "accelerate>=1.13.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "transformers>=5.5.0",
]

local-cu130 = [
    "funasr[llm]>=1.3.1",
    "ilt-editdistance>=0.8.2",
    "sentencepiece>=0.2.1",
    "accelerate>=1.13.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "transformers>=5.5.0",
]

[tool.uv]
exclude-dependencies = ["editdistance"]
conflicts = [
    [{ extra = "local-cpu" }, { extra = "local-cu130" }],
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "local-cpu" },
    { index = "pytorch-cu130", extra = "local-cu130", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]
torchaudio = [
    { index = "pytorch-cpu", extra = "local-cpu" },
    { index = "pytorch-cu130", extra = "local-cu130", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.ruff]
line-length = 88
force-exclude = true
exclude = ["docs"]

[tool.ruff.lint]
extend-select = ["I", "D"]
ignore = ["D100", "D104", "D105", "D107", "D417"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]

[tool.ty.src]
exclude = ["docs"]

[tool.pytest.ini_options]
# Keep the default test flow fast, offline, and non-billable.
addopts = '--strict-markers -m "not integration and not live"'
testpaths = ["tests"]
markers = [
    "integration: tests that exercise cross-subsystem or end-to-end behavior and are excluded from default pytest runs",
    "live: tests that call real external services, require network access and credentials, and may be billable; excluded from default pytest runs",
    "deepl_live: live tests for the DeepL API; requires DEEPL_AUTH_KEY",
    "google_translate_live: live tests for Google Cloud Translation; requires ADC plus GOOGLE_TRANSLATE_PROJECT_ID",
    "openai_live: live tests for the OpenAI API; requires OPENAI_API_KEY",
    "iflytek_live: live tests for the iFLYTEK RTASR API; requires IFLYTEK_APP_ID, IFLYTEK_API_KEY, and IFLYTEK_API_SECRET",
]

[tool.coverage.run]
source = ["vrc_live_caption"]
branch = true
relative_files = true

[tool.coverage.report]
show_missing = true
skip_covered = true
# fail_under = 80
